chore: refactor websocket client
