+++ shared/domain/src/commonMain/kotlin/network/bisq/mobile/client/websocket/ConnectionState.kt
@@ -1,0 +1,7 @@
+package network.bisq.mobile.client.websocket
+
+sealed class ConnectionState {
+    data class Disconnected(val error: Throwable? = null) : ConnectionState()
+    object Connecting : ConnectionState()
+    object Connected : ConnectionState()
+}
